+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Fixed
 
 - Fixed a small bug that would prevent active effects from being created inline on the player character sheet
+- Fixed a small bug that would prevent weapons from being displayed in the cargo tab on the vehicle sheet
 
 ## [v0.20.2]